Update to 1.7.10
Updated to 1.7.10 and included build.gradle

Built on Forge 1.7.10-10.12.3.1230
