+++ README.md
@@ -1,0 +1,9 @@
+# Hangable Maps
+
+「Hangable Maps」はljdp氏が作成したMODです。
+作成したMAPを壁に貼り付けることができます。また、MAPの複製機能も備えています。
+友達とのマルチプレイの時にいいMODはないかな～とJPWikiを眺めていた時に見つけたものの、Forge対応版がなかったので自分で作りました。
+
+Previous releases can be found here:
+
+All existing releases can be found here: https://onedrive.live.com/redir?resid=237C3785E910AA7F!2330&authkey=!APoMgescPI8Xoq8&ithint=folder%2c
