--- src/hide92795/mods/hangablemaps/ServerTickHandler.java
@@ -1,33 +1,0 @@
-package hide92795.mods.hangablemaps;
-
-import java.util.EnumSet;
-
-import cpw.mods.fml.common.ITickHandler;
-import cpw.mods.fml.common.TickType;
-
-public class ServerTickHandler implements ITickHandler {
-	private int nextMapUpdate;
-
-	@Override
-	public void tickStart(EnumSet<TickType> type, Object... tickData) {
-	}
-
-	@Override
-	public void tickEnd(EnumSet<TickType> type, Object... tickData) {
-		++this.nextMapUpdate;
-		if (this.nextMapUpdate > 3) {
-			this.nextMapUpdate = 0;
-			HangableMaps.instance.executeQueue();
-		}
-	}
-
-	@Override
-	public EnumSet<TickType> ticks() {
-		return EnumSet.of(TickType.SERVER);
-	}
-
-	@Override
-	public String getLabel() {
-		return null;
-	}
-}
